requirements.txt
accelerate
bitsandbytes
datasets
pinecone-client[grpc]
sentencepiece
sentence-transformers
torch
transformers
wikipedia
